feat(max-len): broaden ignore list (#130)
Adds strings, template literals, regexp literals, and urls to the
max-len ignore list. All of these things can and sometimes must exceed
this length.

Strings and template-literals are intended for output.
Wrapping/truncation is the job of the terminal.
regexp formulas are not something that can be arbitrarily broken up and
truncated.
urls are strings that should not be broken up.